folders moved to app and added controllers and providers folders
